Create Product and ProductProvider components
The ProductProvider component wraps the Product component and handles figuring out the exact product that is going to be rendered, based on the slug that is passed in. This means the Product component itself only has to worry about rendering the data it's given, not being given more data than it needs, finding the data that it needs to render, and then render it.
This also makes the translation of data to directions a bit clearer (i.e. shoppingBagItems to isInBag)