Showing a product will now retain a search query
This means that when a user clicks on a product, the products displayed in the background won't change, because those products are displayed based on any search query being in the URL.
The issue I was seeing is that when a user opens one of the search results, the products in the background would re-render to display the basic catalogue of showbags. When the user closed the product modal, then the search results would display again. This could be disorientating..